ref(laravel-insights): Split into multiple files
