+++ static/app/views/insights/pages/backend/laravel/issuesWidget.tsx
@@ -1,0 +1,77 @@
+import pick from 'lodash/pick';
+
+import GroupList from 'sentry/components/issues/groupList';
+import {normalizeDateTimeParams} from 'sentry/components/organizations/pageFilters/parse';
+import Panel from 'sentry/components/panels/panel';
+import PanelBody from 'sentry/components/panels/panelBody';
+import {DEFAULT_STATS_PERIOD} from 'sentry/constants';
+import {URL_PARAM} from 'sentry/constants/pageFilters';
+import {t, tct} from 'sentry/locale';
+import {decodeScalar} from 'sentry/utils/queryString';
+import useApi from 'sentry/utils/useApi';
+import {useBreakpoints} from 'sentry/utils/useBreakpoints';
+import {useLocation} from 'sentry/utils/useLocation';
+import useOrganization from 'sentry/utils/useOrganization';
+import {usePageFilterChartParams} from 'sentry/views/insights/pages/backend/laravel/utils';
+import NoGroupsHandler from 'sentry/views/issueList/noGroupsHandler';
+
+export function IssuesWidget({query = ''}: {query?: string}) {
+  const api = useApi();
+  const organization = useOrganization();
+  const location = useLocation();
+
+  const pageFilterChartParams = usePageFilterChartParams({granularity: 'spans-low'});
+
+  const queryParams = {
+    limit: '5',
+    ...normalizeDateTimeParams(
+      pick(location.query, [...Object.values(URL_PARAM), 'cursor'])
+    ),
+    query,
+    sort: 'freq',
+  };
+
+  const breakpoints = useBreakpoints();
+
+  function renderEmptyMessage() {
+    const selectedTimePeriod = location.query.start
+      ? null
+      : // @ts-expect-error TS(7053): Element implicitly has an 'any' type because expre... Remove this comment to see the full error message
+        DEFAULT_RELATIVE_PERIODS[
+          decodeScalar(location.query.statsPeriod, DEFAULT_STATS_PERIOD)
+        ];
+    const displayedPeriod = selectedTimePeriod
+      ? selectedTimePeriod.toLowerCase()
+      : t('given timeframe');
+
+    return (
+      <Panel style={{display: 'flex', justifyContent: 'center', alignItems: 'center'}}>
+        <PanelBody>
+          <NoGroupsHandler
+            api={api}
+            organization={organization}
+            query={query}
+            selectedProjectIds={pageFilterChartParams.project}
+            groupIds={[]}
+            emptyMessage={tct('No [issuesType] issues for the [timePeriod].', {
+              issuesType: '',
+              timePeriod: displayedPeriod,
+            })}
+          />
+        </PanelBody>
+      </Panel>
+    );
+  }
+
+  // TODO(aknaus): Remove GroupList and use StreamGroup directly
+  return (
+    <GroupList
+      orgSlug={organization.slug}
+      queryParams={queryParams}
+      canSelectGroups={false}
+      renderEmptyMessage={renderEmptyMessage}
+      withChart={breakpoints.xlarge}
+      withPagination={false}
+    />
+  );
+}
